refactor: Deleted the Deleted at key at NetworkUser and NetworkVendor
