+++ app/src/main/java/com/carbondev/carboncheck/data/remote/model/NetworkVendor.kt
@@ -11,7 +11,6 @@ data class NetworkVendor(
     val name: String,
     @Json(name = "created_at") val createdAt: Date,
     @Json(name = "updated_at") val updatedAt: Date,
-    @Json(name = "deleted_at") val deletedAt: Date?,
 ) : RemoteMappable<Vendor> {
     companion object {
         const val TABLE_NAME = "vendors"
@@ -22,14 +21,12 @@ data class NetworkVendor(
         const val NAME = "name"
         const val CREATED_AT = "created_at"
         const val UPDATED_AT = "updated_at"
-        const val DELETED_AT = "deleted_at"
 
         val ALL = listOf(
             ID,
             NAME,
             CREATED_AT,
             UPDATED_AT,
-            DELETED_AT,
         )
     }
